pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "fast-jwt-auth"
version = "0.1.4"
authors = [
  { name="romeo-tech", email="romeomanoela18@gmail.com" },
]
description = "Une bibliothèque d'authentification JWT simple pour FastAPI."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Framework :: FastAPI",
]
dependencies = [
    "PyJWT>=2.0.0",
]

[project.urls]
"Homepage" = "https://github.com/PyCoder601/fast_jwt"

[project.optional-dependencies]
test = [
    "pytest",
    "httpx",
    "uvicorn",
]
